# core/api/rest/openapi.yaml
...
openapi: 3.1.0
info:
  title: CypherPanel API
  description: >-
    Operator-facing control-plane API covering Phases 1–3: authentication and
    teams/roles, the server lifecycle, projects/environments, applications and
    the deploy pipeline, managed databases with S3 backups/restore, preview
    environments, notifications, and scheduled tasks. Authentication is a bearer
    session token from /api/v1/auth/login; the GitHub webhook authenticates by
    per-application HMAC instead. Every project-scoped route is authorized by
    team membership (a non-member sees 404, insufficient rank 403).


    Every response — success, error and SSE stream alike — carries an
    `X-Request-Id` header (`components/headers/RequestId`), and every JSON
    error body repeats it as `trace_id`. It is the value to quote in a bug
    report and the key to search for in `GET /api/v1/panel/logs`. Individual
    responses reference the header only where a generated client benefits;
    it is present on all of them.
  version: 0.3.0
servers:
  - url: /
...
paths:
...
  /api/v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: The authenticated account and its teams
      responses:
        "200":
          description: The current user, with the teams they belong to and their role in each.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Set your own display name and timezone (session only)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateProfileRequest" } } }
      responses:
        "200": { description: The updated account., content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/auth/password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change your own password, proving the current one (session only)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ChangePasswordRequest" } } }
      responses:
        "200": { description: Changed; reports how many other sessions were revoked., content: { application/json: { schema: { $ref: "#/components/schemas/ChangePasswordResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { description: The current password is wrong — the same answer a failed sign-in gets., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/auth/me/avatar:
    put:
      operationId: setAvatar
      tags: [auth]
      summary: Replace your own profile photo (session only)
      description: >
        The image is the raw request body. The stored content type is the one
        sniffed from the bytes, never the declared one; PNG, JPEG and WebP are
        accepted and SVG deliberately is not.
      requestBody:
        required: true
        content:
          image/png: { schema: { type: string, format: binary } }
          image/jpeg: { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
      responses:
        "200": { description: Stored., content: { application/json: { schema: { $ref: "#/components/schemas/AvatarResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { description: Larger than the 2 MB limit., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    delete:
      operationId: deleteAvatar
      tags: [auth]
      summary: Remove your own profile photo (session only)
      responses:
        "204": { description: Removed; initials are shown instead. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/v1/auth/sessions/revoke-others:
    post:
      operationId: revokeOtherSessions
      tags: [auth]
      summary: Sign out everywhere except this device
      description: >
        Revokes every session of the caller except the one making the request,
        identified by the presented bearer token.
      responses:
        "200":
          description: How many sessions were revoked.
          content:
            application/json:
              schema:
                type: object
                required: [revoked]
                properties:
                  revoked:
                    type: integer
                    format: int64
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Session token from /api/v1/auth/login.
  headers:
    RequestId:
      description: >-
        Correlation id for this request, minted by the API's outermost
        middleware and present on every response. It is repeated as `trace_id`
        in every JSON error body, and it is the key to search for in
        `GET /api/v1/panel/logs`. An inbound `X-Request-Id` is honoured only
        when the peer is a configured trusted proxy and the value is a short
        printable token; otherwise a fresh id is minted.
      schema: { type: string }
      example: trace_9f3a-11bd-4c02-7e51
  responses:
    BadRequest:
      description: Malformed or invalid request body.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or expired session token.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated but the caller's role is insufficient for this action.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable message; never contains secrets.
        trace_id:
          type: string
          description: >-
            The value of this response's `X-Request-Id` header, repeated in the
            body so a screenshot of a failure carries it. Paste it into a bug
            report or `GET /api/v1/panel/logs` to find the request. Present on
            every response served through the API's middleware.
          example: trace_9f3a-11bd-4c02-7e51
        retry_after_seconds:
          type: integer
          minimum: 1
          description: >-
            Present only on `429`: whole seconds until the throttle lifts, the
            same value as the `Retry-After` header. Count down against it rather
            than guessing.
...
    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: string
          description: usr_… prefixed ID.
        email:
          type: string
        role:
          type: string
        display_name:
          type: string
          description: What teammates see. Empty means the panel falls back to the address.
        timezone:
          type: string
          description: IANA name the UI reads timestamps in. Empty means UTC.
    UpdateProfileRequest:
      type: object
      properties:
        display_name:
          type: string
          maxLength: 64
          description: Empty clears it, and the address is shown instead.
        timezone:
          type: string
          maxLength: 64
          description: An IANA name such as Europe/Berlin. Empty means UTC.
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
          minLength: 8
        revoke_other_sessions:
          type: boolean
          default: false
          description: Sign every other device out. The session making the request survives.
    AvatarResponse:
      type: object
      required: [etag]
      properties:
        etag:
          type: string
          description: Changes when the image does; use it to bust a cached <img>.
...
    ChangePasswordResponse:
      type: object
      required: [revoked]
      properties:
        revoked:
          type: integer
          format: int64
          description: Other sessions signed out; 0 unless asked.
...
    Me:
      type: object
      required: [id, email, role, teams]
      properties:
        id: { type: string }
        email: { type: string }
        role:
          type: string
          description: Panel role.
          enum: [member, admin, owner]
        display_name:
          type: string
          description: What teammates see. Empty means the panel falls back to the address.
        timezone:
          type: string
          description: IANA name the UI reads timestamps in. Empty means UTC.
        teams:
          type: array
          items: { $ref: "#/components/schemas/Team" }
    Team:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id: { type: string, description: tm_… prefixed ID. }
        name: { type: string }
        role:
          type: string
          description: The requesting user's role in this team (present on listings and /auth/me).
          enum: [member, admin, owner]
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
security:
  - bearer: []